+++ grails-app/services/io/xh/hoist/environment/EnvironmentService.groovy
@@ -29,7 +29,6 @@ class EnvironmentService extends BaseService {
 
     void init() {
         _appTimeZone = calcAppTimeZone()
-        super.init();
     }
 
     /**