+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
-## 18.0-SNAPSHOT - under development
+## 18.0-SNAPSHOT
 
 ### 🎁 New Features
+
 
 * Hoist Core v18 provides support for running multi-instance clusters of Hoist application servers.
   Cluster management is provided by the use of Hazelcast (www.hazelcast.com), an open-source library
@@ -41,6 +42,17 @@
   change was needed to better support cross-cluster exception handling. This object is used by
   Hoist internally for catching uncaught exceptions and this change is not expected to impact
   most applications.
+
+* New support for Role Management.
+  * Hoist now supports an out-of-the-box, database-driven system for maintaining a hierarchical set
+    of Roles associating and associating them with individual users.
+  * New system supports app and plug-in specific integrations to AD and other enterprise systems.
+  * Administration of the new system provided by a new admin UI in hoist-react v60 and above.
+
+### ⚙️ Technical
+
+* Add `xh/echoHeaders` utility endpoint. Useful for verifying headers (e.g. `jespa_connection_id`)
+  that are installed by or must pass through multiple ingresses/load balancers.
 
 ### 📚 Libraries
 * grails `6.1.0`